Оптимизация кода по отображению карточек с рецептами и их экарном деталей
