+++ app/src/main/java/com/pidzama/smokecrafthookahapp/presentation/random_generation_recipe/common/PortraitRecipeCard.kt
@@ -1,0 +1,263 @@
+package com.pidzama.smokecrafthookahapp.presentation.random_generation_recipe.common
+
+import android.graphics.Paint
+import android.graphics.Typeface
+import android.util.Log
+import androidx.compose.animation.core.LinearOutSlowInEasing
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.*
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Card
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.StrokeCap
+import androidx.compose.ui.graphics.drawscope.Stroke
+import androidx.compose.ui.graphics.drawscope.rotate
+import androidx.compose.ui.graphics.drawscope.scale
+import androidx.compose.ui.graphics.nativeCanvas
+import androidx.compose.ui.graphics.toArgb
+import androidx.compose.ui.platform.LocalConfiguration
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.pidzama.smokecrafthookahapp.R
+import com.pidzama.smokecrafthookahapp.data.model.RandomRecipeSubList
+import com.pidzama.smokecrafthookahapp.presentation.common.setColorTaste
+import com.pidzama.smokecrafthookahapp.ui.theme.dimens
+
+//карточка рецепта при книжной ориентации
+@Composable
+fun PortraitRecipeCard(
+    input: RandomRecipeSubList,
+    onClickToDetailsScreen: () -> Unit,
+    indexRecipe: Int,
+    listTobaccoWeight: List<Float>,
+    radius: Float
+) {
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = MaterialTheme.dimens.small1)
+            .clickable { onClickToDetailsScreen() },
+        border = BorderStroke(2.dp, color = MaterialTheme.colorScheme.primary),
+        shape = MaterialTheme.shapes.extraLarge,
+        backgroundColor = MaterialTheme.colorScheme.background
+    ) {
+        Row(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(
+                    vertical = MaterialTheme.dimens.small3,
+                    horizontal = MaterialTheme.dimens.small3
+                )
+        ) {
+            Box(
+                modifier = Modifier
+                    .weight(0.5f)
+            ) {
+                LegendRecipePortrait(
+                    input = input,
+                    indexRecipe = indexRecipe,
+                )
+            }
+            Box(
+                modifier = Modifier
+                    .weight(0.5f)
+                    .padding(
+                    ),
+                contentAlignment = Alignment.Center
+            ) {
+                RecipePieChartPortrait(
+                    input = input,
+                    listTobaccoWeight = listTobaccoWeight,
+                    radius = radius
+                )
+            }
+        }
+    }
+}
+
+//легенда ингридиентов при книжной ориентации
+@OptIn(ExperimentalFoundationApi::class)
+@Composable
+fun LegendRecipePortrait(
+    input: RandomRecipeSubList,
+    indexRecipe: Int,
+) {
+    Column(
+        modifier = Modifier
+            .padding(end = MaterialTheme.dimens.small1)
+    ) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = stringResource(id = R.string.recipe),
+                color = MaterialTheme.colorScheme.inverseSurface,
+                style = MaterialTheme.typography.headlineMedium,
+            )
+            Text(
+                text = " №${indexRecipe}",
+                color = MaterialTheme.colorScheme.primary,
+                style = MaterialTheme.typography.headlineMedium,
+            )
+        }
+        input.forEach {
+            Card(
+                modifier = Modifier
+                    .padding(vertical = MaterialTheme.dimens.extraSmall),
+                border = BorderStroke(
+                    width = 2.dp,
+                    color = setColorTaste(it.taste_group)
+                ),
+                shape = MaterialTheme.shapes.medium,
+                backgroundColor = MaterialTheme.colorScheme.background
+            ) {
+                Text(
+                    modifier = Modifier
+                        .padding(
+                            horizontal = MaterialTheme.dimens.small2,
+                            vertical = MaterialTheme.dimens.extraSmall
+                        )
+                        .basicMarquee(),
+                    text = "${it.taste}, ${it.brand}",
+                    style = MaterialTheme.typography.titleLarge,
+                    color = setColorTaste(it.taste_group)
+                )
+            }
+        }
+    }
+}
+
+//диаграмма рецепта при книжной ориентации
+@Composable
+fun RecipePieChartPortrait(
+    input: RandomRecipeSubList,
+    listTobaccoWeight: List<Float>,
+    animDuration: Int = 600,
+    radius: Float,
+) {
+    val density = LocalConfiguration.current.densityDpi
+    var circleCenter by remember {
+        mutableStateOf(Offset.Zero)
+    }
+    val coeff: Float
+    val pieRadius: Float
+    if (density > 400) {
+        coeff = 0.7f
+        pieRadius = radius
+    } else if (density in 261..399) {
+        pieRadius = radius / 1.5f
+        coeff = 1f
+    } else {
+        pieRadius = radius / 2.2f
+        coeff = 1.3f
+    }
+
+    val totalTastyWeight = listTobaccoWeight.sum()
+    val innerRadius = pieRadius - ((MaterialTheme.dimens.small1 * 1.3f).value)
+    var animationPlayed by remember { mutableStateOf(false) }
+    val animateRotation by animateFloatAsState(
+        targetValue = if (animationPlayed) 90f * 12f else 0f, animationSpec = tween(
+            durationMillis = animDuration,
+            delayMillis = 200,
+            easing = LinearOutSlowInEasing
+        )
+    )
+    LaunchedEffect(key1 = true) {
+        animationPlayed = true
+    }
+
+    Box(
+        modifier = Modifier
+            .padding(top = (innerRadius * coeff).dp)
+            .background(Color.Red),
+        contentAlignment = Alignment.Center
+    ) {
+        Canvas(
+            modifier = Modifier
+                .fillMaxWidth()
+                .rotate(animateRotation)
+                .background(Color.White)
+        ) {
+
+            val width = size.width
+            val height = size.height
+            val anglePerValue = 360f / totalTastyWeight
+            var currentStartAngle = 270f
+
+            circleCenter = Offset(x = width / 2, y = height / 2)
+
+            listTobaccoWeight.forEachIndexed { index, tobacco ->
+                val scale = 1.1f
+                val angleToDraw = tobacco * anglePerValue
+
+                scale(scale) {
+                    drawArc(
+                        color = if (input.size >= 3) {
+                            setColorTaste(input[index].taste_group)
+                        } else {
+                            Color.Red
+                        },
+                        startAngle = currentStartAngle,
+                        sweepAngle = angleToDraw,
+                        useCenter = false,
+                        size = Size(
+                            width = pieRadius * 2f,
+                            height = pieRadius * 2f
+                        ),
+                        topLeft = Offset(
+                            x = (width - pieRadius * 2f) / 2f,
+                            y = (height - pieRadius * 2f) / 2f
+                        ),
+                        style = Stroke((pieRadius - innerRadius) * 4f, cap = StrokeCap.Butt)
+                    )
+                    Log.d("MyLog", "width = $width")
+                    Log.d("MyLog", "height = $height")
+                    currentStartAngle += angleToDraw
+                }
+
+                var rotateAngle = currentStartAngle - angleToDraw / 2f - 90f
+                var factor = 1f
+                if (rotateAngle > 90f) {
+                    rotateAngle = (rotateAngle + 180).mod(360f)
+                    factor = -0.92f
+                }
+
+                val percentage =
+                    (tobacco / totalTastyWeight * 100).toInt()
+                drawContext.canvas.nativeCanvas.apply {
+                    if (percentage > 3) {
+                        rotate(rotateAngle) {
+                            drawText(
+                                "$percentage %",
+                                circleCenter.x,
+                                circleCenter.y + (pieRadius + innerRadius / 2f) * factor,
+                                Paint().apply {
+                                    textSize = 15.sp.toPx()
+                                    color = if (input.size >= 3) {
+                                        setColorTaste(input[index].taste_group).toArgb()
+                                    } else {
+                                        Color.Red.toArgb()
+                                    }
+                                    textAlign = Paint.Align.CENTER
+                                    typeface = Typeface.DEFAULT_BOLD
+                                }
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
+
+
