+++ app/src/main/res/values-en/strings.xml
@@ -1,39 +1,41 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">SmokeCraftHookahApp</string>
-    <string name="enter_to_app">Вход в систему</string>
-    <string name="sign_in">Войти</string>
+    <string name="enter_to_app">Welcome</string>
+    <string name="sign_in">Sign in</string>
     <string name="login">Login</string>
     <string name="password">Password</string>
-    <string name="recipe">Рецепт</string>
-    <string name="recipes">Рецепты</string>
-    <string name="cancel">Отменить</string>
-    <string name="write_off">Списать со склада</string>
-    <string name="order">Заказ</string>
-    <string name="orders">Заказы</string>
-    <string name="table">стол</string>
-    <string name="logout">Выйти</string>
-    <string name="account">Аккаунт</string>
-    <string name="storage">Склад</string>
-    <string name="theme">Тема</string>
-    <string name="language">Язык</string>
-    <string name="help">Помощь</string>
-    <string name="ok">Принять</string>
-    <string name="choose_language">Выберите язык</string>
-    <string name="russian">Русский</string>
-    <string name="english">Английский</string>
-    <string name="faq">Вопросы и тех.поддержка</string>
-    <string name="change_language">Изменить язык</string>
-    <string name="change_theme">Изменить темы</string>
-    <string name="info_storage">Информация о продукции</string>
-    <string name="settings_profile">Настройки профиля</string>
-    <string name="profile">Мой профиль</string>
-    <string name="current_orders">Текущие заказы</string>
-    <string name="order_archive">Архив заказов</string>
-    <string name="random_generation">Случайная генерация</string>
-    <string name="branded_mixes">Фирменные миксы</string>
-    <string name="generation_by_ingredient">Генерация по ингредиенту</string>
-    <string name="generate_new_recipe">Сгенерировать новый рецепт</string>
-    <string name="successfully_written_off">Табак успешно списан со склада!</string>
+    <string name="recipe">Recipe</string>
+    <string name="recipes">Recipes</string>
+    <string name="cancel">Cancel</string>
+    <string name="write_off">Write off</string>
+    <string name="order">Order</string>
+    <string name="orders">Orders</string>
+    <string name="table">table</string>
+    <string name="logout">Log out</string>
+    <string name="account">Account</string>
+    <string name="storage">Storage</string>
+    <string name="theme">Theme</string>
+    <string name="language">Language</string>
+    <string name="help">Help</string>
+    <string name="ok">OK</string>
+    <string name="choose_language">Choose language</string>
+    <string name="russian">Russian</string>
+    <string name="english">English</string>
+    <string name="faq">Questions and technical support</string>
+    <string name="change_language">Change application language</string>
+    <string name="change_theme">Change application theme</string>
+    <string name="info_storage">Product Information</string>
+    <string name="settings_profile">Settings your profile</string>
+    <string name="profile">My profile</string>
+    <string name="current_orders">Current orders</string>
+    <string name="order_archive">Order archive</string>
+    <string name="random_generation">Random generation</string>
+    <string name="branded_mixes">Branded mixes</string>
+    <string name="generation_by_ingredient">Generation by ingredient</string>
+    <string name="generate_new_recipe">Generate new recipe</string>
+    <string name="successfully_written_off">Tobacco was successfully written off from the warehouse</string>
+
+
 
 </resources>
